ido: QCamera2: Fix manual ISO control
Our proprietary camera libraries were not built by the OEM with manual
ISO in mind, so there are many strange requirements that need to be in
place in order for a manual ISO setting to stick. These weird
requirements are all located in the backend, so this patch involves
sending multiple commands to the backend in order to attain the
special condition where manual ISO settings are honored.

Change-Id: I2c21082ac51f0dc8d2dfcd0a3a9ebafdcef12e2b
(cherry picked from commit 7d2c918bba97bbe997ebd41f589e3e62f601b915)